requirements.txt
python-telegram-bot[socks]==20.7
openai==1.12.0
xai-sdk==1.6.1; python_version >= "3.10"
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx[socks]==0.25.2
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
cos-python-sdk-v5==1.9.41
